Consolidate permissions e2e specs (#43552)
