--- e2e/test/scenarios/permissions/create-queries/no.cy.spec.js
@@ -1,48 +1,0 @@
-import {
-  assertPermissionTable,
-  selectSidebarItem,
-  modal,
-  restore,
-  modifyPermission,
-} from "e2e/support/helpers";
-
-const NATIVE_QUERIES_PERMISSION_INDEX = 0;
-
-describe("scenarios > admin > permissions > create queries > no", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-  });
-
-  it("should allow setting create queries to 'no' in group view", () => {
-    cy.visit("/admin/permissions/data");
-
-    selectSidebarItem("data");
-
-    modifyPermission("Sample Database", NATIVE_QUERIES_PERMISSION_INDEX, "No");
-
-    assertPermissionTable([["Sample Database", "No"]]);
-
-    cy.button("Save changes").click();
-
-    modal().within(() => {
-      cy.findByText("Save permissions?");
-      cy.button("Yes").click();
-    });
-
-    assertPermissionTable([["Sample Database", "No"]]);
-
-    cy.findByTextEnsureVisible("Sample Database").click();
-
-    assertPermissionTable([
-      ["Accounts", "No"],
-      ["Analytic Events", "No"],
-      ["Feedback", "No"],
-      ["Invoices", "No"],
-      ["Orders", "No"],
-      ["People", "No"],
-      ["Products", "No"],
-      ["Reviews", "No"],
-    ]);
-  });
-});
